@@ -1,144 +1,76 @@
 {
-  "name": "eai-schema-toolkit",
-  "version": "1.0.0",
-  "description": "EAI Schema Toolkit - Enterprise Application Integration Schema Conversion and Validation Tool",
-  "homepage": "https://yakdoli.github.io/eai-schema/",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/yakdoli/eai-schema.git"
-  },
-  "bugs": {
-    "url": "https://github.com/yakdoli/eai-schema/issues"
-  },
-  "keywords": [
-    "eai",
-    "schema",
-    "xml",
-    "json",
-    "yaml",
-    "conversion",
-    "validation",
-    "enterprise"
-  ],
-  "main": "dist/index.js",
+  "name": "eai-work-tool-clone",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "build": "npx tsc",
-    "dev": "tsx watch src/index.ts",
-    "start": "node dist/index.js",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
-    "test:integration": "jest --testPathPattern=integration",
-    "test:unit": "jest --testPathPattern=services",
-    "test:e2e": "playwright test",
-    "test:e2e:headed": "playwright test --headed",
-    "test:performance": "jest --testPathPattern=performance --runInBand --detectOpenHandles",
-    "test:load": "artillery run src/__tests__/performance/load-test.config.yml",
-    "test:all-integration": "npm run test:integration && npm run test:e2e && npm run test:performance",
-    "lint": "eslint src",
-    "lint:fix": "eslint src --fix",
-    "format": "prettier --write \"src/**/*.{ts,js,json}\"",
-    "format:check": "prettier --check \"src/**/*.{ts,js,json}\"",
-    "type-check": "tsc --noEmit",
-    "clean": "rimraf dist",
-    "prebuild": "npm run clean",
-    "postbuild": "echo \"Build completed successfully\"",
-    "deploy": "npm run build && gh-pages -d docs",
-    "predeploy": "npm run build",
-    "heroku-prebuild": "echo \"Prebuild script executed\"",
-    "heroku-postbuild": "npm run build",
-    "test:all": "./scripts/test-runner.sh",
-    "security:audit": "npm audit --audit-level=moderate",
-    "security:fix": "npm audit fix",
-    "security:licenses": "license-checker --summary",
-    "release": "semantic-release",
-    "release:dry": "semantic-release --dry-run",
-    "ci:validate": "npm run type-check && npm run lint && npm run format:check",
-    "ci:test": "npm run test:unit && npm run test:integration",
-    "ci:build": "npm run build",
-    "mcp:start": "cd mcp-server && npm start",
-    "mcp:install": "cd mcp-server && npm install"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@types/bcrypt": "^5.0.2",
-    "@types/express-validator": "^2.20.33",
-    "@types/handsontable": "^0.31.2",
-    "@types/js-yaml": "^4.0.9",
-    "@types/jsonwebtoken": "^9.0.10",
-    "@types/swagger-jsdoc": "^6.0.4",
-    "@types/swagger-ui-express": "^4.1.6",
-    "@types/xml2js": "^0.4.14",
-    "ajv": "^8.17.1",
-    "ajv-formats": "^2.1.1",
-    "bcrypt": "^5.1.1",
-    "compression": "^1.7.4",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.5",
-    "express": "^4.21.1",
-    "express-rate-limit": "^7.4.0",
-    "express-validator": "^7.2.1",
-    "fast-xml-parser": "^4.4.1",
-    "handsontable": "^16.0.1",
-    "helmet": "^7.1.0",
-    "js-yaml": "^4.1.0",
-    "jsonwebtoken": "^9.0.2",
-    "libxmljs2": "^0.37.0",
-    "morgan": "^1.10.0",
-    "multer": "^1.4.5-lts.1",
-    "prom-client": "^15.1.3",
-    "prometheus-gc-stats": "^1.1.0",
-    "socket.io": "^4.8.1",
-    "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.1",
-    "winston": "^3.15.0",
-    "xml-js": "^1.6.11",
-    "xml2js": "^0.6.2",
-    "xmldom": "^0.6.0",
-    "yaml-validator": "^5.1.0",
-    "zod": "^3.23.8"
+    "@hookform/resolvers": "^5.0.1",
+    "@radix-ui/react-accordion": "^1.2.10",
+    "@radix-ui/react-alert-dialog": "^1.1.13",
+    "@radix-ui/react-aspect-ratio": "^1.1.6",
+    "@radix-ui/react-avatar": "^1.1.9",
+    "@radix-ui/react-checkbox": "^1.3.1",
+    "@radix-ui/react-collapsible": "^1.1.10",
+    "@radix-ui/react-context-menu": "^2.2.14",
+    "@radix-ui/react-dialog": "^1.1.13",
+    "@radix-ui/react-dropdown-menu": "^2.1.14",
+    "@radix-ui/react-hover-card": "^1.1.13",
+    "@radix-ui/react-label": "^2.1.6",
+    "@radix-ui/react-menubar": "^1.1.14",
+    "@radix-ui/react-navigation-menu": "^1.2.12",
+    "@radix-ui/react-popover": "^1.1.13",
+    "@radix-ui/react-progress": "^1.1.6",
+    "@radix-ui/react-radio-group": "^1.3.6",
+    "@radix-ui/react-scroll-area": "^1.2.8",
+    "@radix-ui/react-select": "^2.2.4",
+    "@radix-ui/react-separator": "^1.1.6",
+    "@radix-ui/react-slider": "^1.3.4",
+    "@radix-ui/react-slot": "^1.2.2",
+    "@radix-ui/react-switch": "^1.2.4",
+    "@radix-ui/react-tabs": "^1.1.11",
+    "@radix-ui/react-toggle": "^1.1.8",
+    "@radix-ui/react-toggle-group": "^1.1.9",
+    "@radix-ui/react-tooltip": "^1.2.6",
+    "@tailwindcss/vite": "^4.1.7",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "^1.1.1",
+    "date-fns": "^4.1.0",
+    "embla-carousel-react": "^8.6.0",
+    "framer-motion": "^12.15.0",
+    "input-otp": "^1.4.2",
+    "lucide-react": "^0.510.0",
+    "next-themes": "^0.4.6",
+    "react": "^19.1.0",
+    "react-day-picker": "8.10.1",
+    "react-dom": "^19.1.0",
+    "react-hook-form": "^7.56.3",
+    "react-resizable-panels": "^3.0.2",
+    "react-router-dom": "^7.6.1",
+    "recharts": "^2.15.3",
+    "sonner": "^2.0.3",
+    "tailwind-merge": "^3.3.0",
+    "tailwindcss": "^4.1.7",
+    "vaul": "^1.1.2",
+    "zod": "^3.24.4"
   },
   "devDependencies": {
-    "@eslint/js": "^9.35.0",
-    "@playwright/test": "^1.55.0",
-    "@semantic-release/changelog": "^6.0.3",
-    "@semantic-release/commit-analyzer": "^13.0.0",
-    "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^11.0.0",
-    "@semantic-release/npm": "^12.0.1",
-    "@semantic-release/release-notes-generator": "^14.0.1",
-    "@types/compression": "^1.7.4",
-    "@types/cors": "^2.8.16",
-    "@types/express": "^4.17.21",
-    "@types/jest": "^29.5.14",
-    "@types/morgan": "^1.9.7",
-    "@types/multer": "^1.4.11",
-    "@types/node": "^22.9.0",
-    "@types/socket.io": "^3.0.1",
-    "@types/supertest": "^6.0.1",
-    "@typescript-eslint/eslint-plugin": "^8.42.0",
-    "@typescript-eslint/parser": "^8.42.0",
-    "artillery": "^2.0.24",
-    "autocannon": "^8.0.0",
-    "clinic": "^13.0.0",
-    "conventional-changelog-conventionalcommits": "^8.0.0",
-    "eslint": "^9.15.0",
-    "gh-pages": "^6.1.1",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^30.1.2",
-    "license-checker": "^25.0.1",
-    "playwright": "^1.55.0",
-    "prettier": "^3.4.2",
-    "rimraf": "^5.0.10",
-    "semantic-release": "^24.2.0",
-    "socket.io-client": "^4.8.1",
-    "supertest": "^6.3.4",
-    "ts-jest": "^29.4.1",
-    "tsx": "^4.19.1",
-    "typescript": "^5.7.2",
-    "typescript-eslint": "^8.42.0"
+    "@eslint/js": "^9.25.0",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "@vitejs/plugin-react": "^4.4.1",
+    "eslint": "^9.25.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^16.0.0",
+    "tw-animate-css": "^1.2.9",
+    "vite": "^6.3.5"
   },
-  "engines": {
-    "node": ">=22.0.0",
-    "npm": ">=10.0.0"
-  }
+  "packageManager": "pnpm@10.4.1+sha512.c753b6c3ad7afa13af388fa6d808035a008e30ea9993f58c6663e2bc5ff21679aa834db094987129aa4d488b86df57f7b634981b2f827cdcacc698cc0cfb88af"
 }
