@@ -1,114 +1,33 @@
-const js = require('@eslint/js');
-const tseslint = require('@typescript-eslint/eslint-plugin');
-const tsparser = require('@typescript-eslint/parser');
+import js from '@eslint/js'
+import globals from 'globals'
+import reactHooks from 'eslint-plugin-react-hooks'
+import reactRefresh from 'eslint-plugin-react-refresh'
 
-module.exports = [
-  js.configs.recommended,
+export default [
+  { ignores: ['dist'] },
   {
-    // JavaScript 파일용 설정
-    files: ['**/*.js'],
+    files: ['**/*.{js,jsx}'],
     languageOptions: {
-      ecmaVersion: 'latest',
-      sourceType: 'module',
-      globals: {
-        console: 'readonly',
-        process: 'readonly',
-        Buffer: 'readonly',
-        __dirname: 'readonly',
-        __filename: 'readonly',
-        global: 'readonly',
-        module: 'readonly',
-        require: 'readonly',
-        exports: 'readonly',
-        express: 'readonly'
-      }
-    },
-    rules: {
-      'no-unused-vars': ['error', { 
-        argsIgnorePattern: '^_',
-        varsIgnorePattern: '^_' 
-      }],
-      'no-console': 'warn'
-    }
-  },
-  {
-    files: ['**/*.ts', '**/*.tsx'],
-    languageOptions: {
-      parser: tsparser,
+      ecmaVersion: 2020,
+      globals: globals.browser,
       parserOptions: {
         ecmaVersion: 'latest',
-        sourceType: 'module'
+        ecmaFeatures: { jsx: true },
+        sourceType: 'module',
       },
-      globals: {
-        console: 'readonly',
-        process: 'readonly',
-        Buffer: 'readonly',
-        __dirname: 'readonly',
-        __filename: 'readonly',
-        global: 'readonly',
-        module: 'readonly',
-        require: 'readonly',
-        exports: 'readonly'
-      }
     },
     plugins: {
-      '@typescript-eslint': tseslint
+      'react-hooks': reactHooks,
+      'react-refresh': reactRefresh,
     },
     rules: {
-      // TypeScript 관련 규칙
-      '@typescript-eslint/no-unused-vars': ['error', { 
-        argsIgnorePattern: '^_',
-        varsIgnorePattern: '^_',
-        caughtErrorsIgnorePattern: '^_'
-      }],
-      '@typescript-eslint/no-explicit-any': 'off', // 임시로 비활성화
-      '@typescript-eslint/explicit-function-return-type': 'off',
-      '@typescript-eslint/explicit-module-boundary-types': 'off',
-      '@typescript-eslint/no-inferrable-types': 'warn', // error에서 warn으로 완화
-      
-      // 일반 JavaScript 규칙 비활성화 (TypeScript에서 처리)
-      'no-unused-vars': 'off',
-      'no-undef': 'off',
-      
-      // 코드 품질 규칙
-      'prefer-const': 'error',
-      'no-var': 'error',
-      'no-console': 'warn',
-      'eqeqeq': 'error',
-      'curly': 'error',
-      'no-case-declarations': 'error',
-      'no-useless-escape': 'error',
-      'no-unreachable': 'error'
-    }
-  },
-  {
-    files: ['**/*.test.ts', '**/*.spec.ts'],
-    languageOptions: {
-      globals: {
-        describe: 'readonly',
-        it: 'readonly',
-        test: 'readonly',
-        expect: 'readonly',
-        beforeEach: 'readonly',
-        afterEach: 'readonly',
-        beforeAll: 'readonly',
-        afterAll: 'readonly',
-        jest: 'readonly'
-      }
+      ...js.configs.recommended.rules,
+      ...reactHooks.configs.recommended.rules,
+      'no-unused-vars': ['error', { varsIgnorePattern: '^[A-Z_]' }],
+      'react-refresh/only-export-components': [
+        'warn',
+        { allowConstantExport: true },
+      ],
     },
-    rules: {
-      '@typescript-eslint/no-explicit-any': 'off',
-      'no-console': 'off'
-    }
   },
-  {
-    ignores: [
-      'dist/**',
-      'node_modules/**',
-      'coverage/**',
-      'docs/**',
-      '*.js',
-      '!eslint.config.js'
-    ]
-  }
-];
+]
